chore: refactory styles and use polished to have proper colors
